moved methods to module so they're easier to test, added specs for auth method
